Additions and modifications
In this commit, I created the html page that Prof can view and see students who are not currently enrolled in project yet.
I also changed some of the profs text fields to area text to allow more characters to be written. A button was added to Prof to allow them to view unenrolled students.